src/multi_class_text_classifier/tensorflow/requirements.txt
numpy==1.23.4
pandas==1.5.1
PyYAML==5.4.1
scikit_learn==1.1.3
tensorflow_macos==2.10.0
tqdm==4.64.1
transformers==4.24.0
